Code Restructuring and forEach method implementation for functional interface
